testing generalizations to 3+ legged seastars

+++ PPO_ants.py
@@ -217,8 +217,8 @@ def main():
             
         # logging
         if i_episode % log_interval == 0:
-            avg_length = int(avg_length/log_interval)
-            running_reward = int((running_reward/log_interval))
+            avg_length = avg_length/log_interval
+            running_reward = running_reward/log_interval
             
             print('Episode {} \t Avg length: {} \t Avg reward: {}'.format(i_episode, avg_length, running_reward))
             running_reward = 0
